requirements.txt
python-dotenv
openai==0.28
requests
flask
spotify
gradio